put the user-related js stuff into a separate folder and now the app doesnt bug while launching
