+++ srcs/front/js/views/Signup.js
@@ -1,5 +1,6 @@
 import CONFIG from "../config.js";
 
+
 import AbstractView from "./AbstractView.js";
 export default class extends AbstractView {
     constructor() {
